requirements.txt
azureml-pipeline==1.0.*
python-dotenv==0.10.3